finished bidService and updated AucitonStatus enum and updated User and Item Services

+++ src/main/java/com/example/lusterz/auction_house/service/AuctionItemService.java
@@ -1,5 +1,6 @@
 package com.example.lusterz.auction_house.service;
 
+import java.time.LocalDateTime;
 import java.util.List;
 
 import org.springframework.stereotype.Service;
@@ -10,8 +11,11 @@ import com.example.lusterz.auction_house.exception.UserException;
 import com.example.lusterz.auction_house.model.AuctionItem;
 import com.example.lusterz.auction_house.model.Bid;
 import com.example.lusterz.auction_house.model.User;
+import com.example.lusterz.auction_house.model.enums.AuctionStatus;
 import com.example.lusterz.auction_house.repository.AuctionItemRepository;
 import com.example.lusterz.auction_house.repository.UserRepository;
+
+import jakarta.transaction.Transactional;
 
 @Service
 public class AuctionItemService {
@@ -39,9 +43,14 @@ public class AuctionItemService {
                 .orElseThrow(() -> AuctionItemException.NotFound.byId(itemId));
     }
     
+    @Transactional
     public AuctionItem createItem(AuctionItemRequest auctionItemRequest) {
         User seller = userRepository.findById(auctionItemRequest.sellerId())
             .orElseThrow(() -> UserException.NotFound.byId(auctionItemRequest.sellerId()));
+
+        if (auctionItemRequest.endTime().isAfter(LocalDateTime.now().plusWeeks(4))) {
+            throw AuctionItemException.InvalidState.invalidDuration();
+        }
 
         AuctionItem newItem = new AuctionItem();
 
@@ -51,11 +60,13 @@ public class AuctionItemService {
         newItem.setStartingPrice(auctionItemRequest.startingPrice());
         newItem.setStartTime(auctionItemRequest.startTime());
         newItem.setEndTime(auctionItemRequest.endTime());
+        
         newItem.setSeller(seller);
 
         return itemRepository.save(newItem);
     }
 
+    @Transactional
     public AuctionItem updateItem(Long itemId, Long userId, AuctionItemRequest auctionItemRequest) {
         AuctionItem updatedItem = itemRepository.findById(itemId)
             .orElseThrow(() -> AuctionItemException.NotFound.byId(itemId));
@@ -68,6 +79,10 @@ public class AuctionItemService {
             throw AuctionItemException.InvalidState.hasBids();
         }
 
+        if (updatedItem.getEndTime().isAfter(LocalDateTime.now().plusWeeks(4))) {
+            throw AuctionItemException.InvalidState.invalidDuration();
+        }
+
         updatedItem.setTitle(auctionItemRequest.title());
         updatedItem.setDescription(auctionItemRequest.description());
         updatedItem.setItemImageUrl(auctionItemRequest.itemImageUrl());
@@ -78,6 +93,7 @@ public class AuctionItemService {
         return itemRepository.save(updatedItem);
     }
 
+    @Transactional
     public void deleteItem(Long itemId, Long userId) {
         AuctionItem deletedItem = itemRepository.findById(itemId)
             .orElseThrow(() -> AuctionItemException.NotFound.byId(itemId));
@@ -89,6 +105,11 @@ public class AuctionItemService {
         if (!deletedItem.getBidHistory().isEmpty()) {
             throw AuctionItemException.InvalidState.hasBids();
         }
+
+        if (!deletedItem.getStatus().equals(AuctionStatus.PENDING)) {
+            throw AuctionItemException.InvalidState.alreadyStarted();
+        }
+
         itemRepository.delete(deletedItem);
     }
 }
